WIP: refactor

+++ pid/IMU.java
@@ -4,7 +4,7 @@ package org.redshiftrobotics.lib.pid;
  * Created by adam on 9/16/17.
  */
 public interface IMU {
-    float getAngularRotationX();
-    float getAngularRotationY();
-    float getAngularRotationZ();
+    double getAngularRotationX();
+    double getAngularRotationY();
+    double getAngularRotationZ();
 }
